# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Demo",
        "contact": {
            "email": "martin.cincurak@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/projects/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get project information",
                "description": "Returns project data",
                "operationId": "getProjectById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Log in user",
                "description": "Log in user",
                "operationId": "loginUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "post log in",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "post log in",
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": 10,
                                    "password": "Jessica Smith"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "User does not exist"
                    },
                    "200": {
                        "description": "Return Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Token": {
                                            "description": "post log in",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "Token": "wrfdfgrt"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}